Create OnboardingView, add app icons
Create OnboardingView. There was an issue with handling Image().resizable() support — in this implementation the image is displayed correctly on the device which was used in the design template — iPhone 12 mini. When run on the devices with other screen dimensions the image becomes stretched. The appropriate way to handle image resizing preserving correct displaying of other UI elements at the same time wasn't found.
Here and forth, UI supports all screens with the @2x and higher scale factors. The only device @1x which supports iOS 14.0 and higher is iPhone SE (1st generation), so it was decided to ignore the support of @1x screens.