Add SwiftLint, fix problems of code style
Install SwiftLint as a dependency to the project. The only disabled rule of the linter for now is "trailing_whitespace" (but it can be considered to disable the "line_length" rule as well).
Fix the project warnings and errors which are caused by the code style violations.

Also, start specifying the exact versions of the dependencies in Podfile for better control over the project stability.