Status bar style (default mode and modifier)
Set status bar default style to be white for any color scheme (see the changes in Info.plist), since almost all of the views in the app will have a dark background.

Also, create a modifier to be able to selectively change status bar style for those views which have light background (OnboardingView is one of them). The modifier uses deprecated 'UIApplication.shared.statusBarStyle' property but it will be fixed in future (hope that SwiftUI will get a native way to set status bar style to the moment it all crashes).