AuthField/DataField, PhoneNumberField, RenteeTextFieldStyle, Corner
Make some refactoring with AuthField. In the first place — rename it to DataField and move it to the "Shared" folder since it will be reused in future views. Also, rename the LookUpSecureField to SecureLookUpField for better name consistency with SecureField

Create PhoneNumberField. The specific about the field is that it will have the capability to determine the country of the phone number by analyzing the country code phone prefix (e.g. +380). Analyzing will be performed in real-time, so we need to pass a binding with the flag emoji from the business layer. It is done by passing the binding when specifying the Kind.phoneNumber as an associated value when initializing the DataField.

Since the UI for the phone number field has to display the country flag, we need some special behavior when assigning text field style, namely — we need to round fewer corners because we will attach the view which will contain our flag emoji (we want to merge them invisible as if it was a whole text field). For this purpose, now use not the static var TextFieldStyle.rentee but use the static func TextFieldStyle.rentee(corners:). This allows us to choose which corner needs to be round in the runtime.

To do this, create a modifier cornerRadius(_:corners:) which will look naturally in the SwiftUI system and will intuitively associate with the native cornerRadius(_:antialised). Under the hood, it uses RoundedCornersRectangle shape, which has some bridging behavior between UIKit and SwiftUI incapsulated.

Create a custom enum Corner and structure Corner.Set for sake of using it in the TextFieldStyle.rentee(corners:), cornerRadius(_:corners:) and RoundedCornersRectangle.