MFACodeForm, MFACodeField, AuthorizationView management foundation
Create MFACodeForm which contains UI for phone code verification. There are a number of conceptual changes in forms:
- Now all forms encapsulate all of the content displayed on the authorization sheet;
- For interacting with button events use modifiers prefixed with 'on' and taking closure as its argument, e.g. onResend { .. } ;
- Apply size-dependant button styles on the most top-level, if possible;
- Store all forms in the "Forms" folder.
SignUpForm was updated accordingly to conform to these conceptual requirements.

Create MFACodeField view. Since the functionality it requires is quite specific the best solution for it would be to create the field using UIKit and then port it into SwiftUI via UIViewRepresentable.

Write a sample code which will show the transition animation between SignUpForm and MFACodeForm. Press "Sign Up" or "Resend" buttons to switch between adjacent views