Start using @Environment to pass button size (OnboardingView)
Start using @Environment to pass button size. This is much flexible way for passing some predefined value common for all subviews of a parent view and much more native. Also, if we specify button size on the most top-level available, when reading code on local level it's confusing to tell from where does the button get its style. So applying button style localy is much more straigtforward and intuitive understandable way of dealing with it. Here and further use this approach for specifying button styles.

Apply this change to all related files in "OnboardingView" folder.